refactor(auction): bookmark로 통일
