--- src/main/java/com/bugzero/rarego/boundedContext/auction/in/dto/WishlistListResponseDto.java
@@ -1,14 +1,0 @@
-package com.bugzero.rarego.boundedContext.auction.in.dto;
-
-import com.bugzero.rarego.shared.auction.dto.AuctionListResponseDto;
-import lombok.Builder;
-
-@Builder
-public record WishlistListResponseDto(
-        Long bookmarkId,
-        AuctionListResponseDto auctionInfo
-) {
-    public static WishlistListResponseDto of(Long bookmarkId, AuctionListResponseDto auctionInfo) {
-        return new WishlistListResponseDto(bookmarkId, auctionInfo);
-    }
-}
