+++ src/main/java/com/bugzero/rarego/bounded_context/auction/in/dto/AuctionAddBookmarkResponseDto.java
@@ -1,0 +1,10 @@
+package com.bugzero.rarego.bounded_context.auction.in.dto;
+
+public record AuctionAddBookmarkResponseDto(
+        boolean bookmarked,
+        Long auctionId
+) {
+    public static AuctionAddBookmarkResponseDto of(boolean bookmarked, Long auctionId) {
+        return new AuctionAddBookmarkResponseDto(bookmarked, auctionId);
+    }
+}
