+++ src/main/java/com/bugzero/rarego/bounded_context/auction/in/dto/AuctionBookmarkListResponseDto.java
@@ -1,0 +1,14 @@
+package com.bugzero.rarego.bounded_context.auction.in.dto;
+
+import com.bugzero.rarego.shared.auction.dto.AuctionListResponseDto;
+import lombok.Builder;
+
+@Builder
+public record AuctionBookmarkListResponseDto(
+        Long bookmarkId,
+        AuctionListResponseDto auctionInfo
+) {
+    public static AuctionBookmarkListResponseDto of(Long bookmarkId, AuctionListResponseDto auctionInfo) {
+        return new AuctionBookmarkListResponseDto(bookmarkId, auctionInfo);
+    }
+}
